pyproject.toml
[project]
name = "douleme-vision-engine"
version = "0.1.0"
description = "Python vision engine prototype for low-grid bead-art generation"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1",
  "uvicorn>=0.34,<1",
  "numpy>=2.0,<3",
  "pillow>=10.0,<12",
  "opencv-python-headless>=4.10,<5",
  "scikit-image>=0.24,<1",
  "scipy>=1.14,<2",
  "pydantic>=2.9,<3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["python/vision_engine"]

